requirements.txt
allure_python_commons==2.9.43
Faker==19.6.1
loguru==0.6.0
PyMySQL==1.0.2
pytest==7.3.1
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
